Move Status enum and rename file
